requirements.txt
aiohttp
pycryptodomex
hashlib